Simplify api interface
