+++ aioredis/__init__.py
@@ -27,8 +27,6 @@ from .errors import (
     SlaveReplyError,
     )
 
-from .streams import ReadStream, ReadGroupStream
-
 __version__ = '1.1.0'
 
 __all__ = [
